+++ CHANGELOG.md
@@ -36,6 +36,13 @@ All notable changes to this project will be documented in this file. This projec
 - The `Contracts\Schema\Container::resources()` method has been removed, in favour of resource factories using the
   schema container's `existsForModel()` and `schemaForModel()` methods.
 
+### Fixed
+
+- The base `Server` class now correctly passes extra parameters in its `url()` method. Previously these were passed to
+  Laravel's `url()` helper - but that helper only appends extra parameters if there is no HTTP host in the provided
+  path. The server's `url()` method now passes these as we *always* went them appended, regardless of whether the API's
+  base path has a HTTP host or not.
+
 ## [1.0.0-beta.1] - 2021-03-30
 
 ### Added